requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
filelock==3.16.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.5
